Add view command, embedding, and caching
This is mostly related to embedding pastes in external pages. Script
tags can be used to link an existing paste, or format a new paste based
on the text enclosed in the script tag. Code tags can be converted to
pastes as well.

When formatting text that is not in an existing paste, the text is not
saved and is not shown on the main page. This is done using the view
command.

Additionally, caching was added to improve load times, specifically for
text enclosed in script/code tags. The TOhtml output is saved in a file
based on the md5sum of the text and some system version information.

A new response function was also added to simplify some processing. It
combines the message and header functions and supports additional things
like gzipped input files (for cached pages) and Access Control headers.